+++ src/app/common/f-chip/chip.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'f-chip',
-  templateUrl: './f-chip.component.html',
-  styleUrls: ['./f-chip.component.scss'],
+  templateUrl: './chip.component.html',
+  styleUrls: ['./chip.component.scss'],
 })
 export class FChipComponent {}
