+++ src/app/units/states/tasks/viewer/directives/task-sheet-view/task-sheet-view.component.ts
@@ -3,8 +3,8 @@ import { TaskDefinition } from 'src/app/api/models/task-definition';
 
 @Component({
   selector: 'f-task-sheet-view',
-  templateUrl: './f-task-sheet-view.component.html',
-  styleUrls: ['./f-task-sheet-view.component.scss'],
+  templateUrl: './task-sheet-view.component.html',
+  styleUrls: ['./task-sheet-view.component.scss'],
 })
 export class FTaskSheetViewComponent {
   @Input() taskDef: TaskDefinition;
